[TASK] Refactoring for version 3.0
Some code cleaning

+++ Classes/Domain/Validator/UniqueValidator.php
@@ -33,7 +33,7 @@ class UniqueValidator extends AbstractValidator
             return $this->isValidState();
         }
         foreach ($this->settings['validation']['unique'] as $marker => $amount) {
-            if ((int) $amount === 0) {
+            if ((int)$amount === 0) {
                 continue;
             }
             foreach ($mail->getAnswers() as $answer) {
@@ -63,12 +63,12 @@ class UniqueValidator extends AbstractValidator
      */
     protected function getStoragePid()
     {
-        $pid = (int) $this->settings['main']['pid'];
+        $pid = (int)$this->settings['main']['pid'];
         if (!empty($this->flexForm['settings']['flexform']['main']['pid'])) {
-            $pid = (int) $this->flexForm['settings']['flexform']['main']['pid'];
+            $pid = (int)$this->flexForm['settings']['flexform']['main']['pid'];
         }
         if ($pid === 0) {
-            $pid = (int) FrontendUtility::getCurrentPageIdentifier();
+            $pid = FrontendUtility::getCurrentPageIdentifier();
         }
         return $pid;
     }
